data received/ chagne tempkf from int to float/change style to material/insert constructor to weatherlistimpl

+++ app/src/main/java/com/example/tnweather/model/Main.java
@@ -18,7 +18,7 @@ public class Main implements Serializable {
 	private double grndLevel;
 
 	@SerializedName("temp_kf")
-	private int tempKf;
+	private float tempKf;
 
 	@SerializedName("humidity")
 	private int humidity;
@@ -60,7 +60,7 @@ public class Main implements Serializable {
 		this.tempKf = tempKf;
 	}
 
-	public int getTempKf(){
+	public float getTempKf(){
 		return tempKf;
 	}
